pyproject.toml
[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dns-agent"
version = "0.1.0"
description = "DNS sinkhole server with ad/malware blocking, caching, database logging, and web dashboard"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "DNS Agent"}
]
keywords = ["dns", "sinkhole", "ad-blocker", "privacy", "security"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: System Administrators",
    "Topic :: Internet :: Name Service (DNS)",
    "Topic :: System :: Networking",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "colorama>=0.4.6",
    "dnspython>=2.8.0",
    "flask>=3.1.0",
    "loguru>=0.7.3",
    "pyyaml>=6.0.3",
]

[project.scripts]
dnsagent = "main:main"

[project.urls]
Homepage = "https://github.com/yourusername/dns-agent"
Repository = "https://github.com/yourusername/dns-agent"

[tool.setuptools]
py-modules = ["main"]

[tool.setuptools.packages.find]
include = ["core*", "ui*"]

[tool.setuptools.package-data]
"*" = ["*.yml", "*.yaml", "*.html", "*.css", "*.js"]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "pytest>=9.0.2",
    "ruff>=0.14.10",
]
